Remove useResizeEvent hook
It turns out that you can only receive resize events for the window
and not for elements other than in Internet Explorer. This makes
this hook useless. I've hacked around the problem by going back to
resizing every frame, but I now debounce the actual resize so that
we don't do it while resizing actively. It's important that the
debounce is inside the check against the current size because if
we actually hit the debounce call every frame even when not resizing,
we would never actually resize at all because we would always defer
the timer for another round.