Refactor cube and line into common Mesh class
There was a lot of duplication across both of these things and I'm
likely to continue using a similar overall approach to all things I want
to render, so it makes sense to distill the common components into a
class that can be shared. The remaining code in the cube and line
classes is much more focused on what is unique about these things rather
than having a lot of webgl code mixed in.